+++ src/services/game_state.rs
@@ -4,7 +4,7 @@ pub mod player;
 
 use super::map;
 use super::messenger;
-use super::minecraft_protocol;
+use super::minecraft_types;
 use super::packet;
 use super::packet_handlers::gameplay_router;
 use super::packet_processor;
